+++ src/apis/http/index.js
@@ -2,8 +2,18 @@ import bank from './bank';
 import commercio from './commercio';
 import distribution from './distribution';
 import gaiaRest from './gaiaRest';
+import keybase from './keybase';
 import staking from './staking';
 import tendermintRpc from './tendermintRpc';
 import tx from './tx';
 
-export { bank, commercio, distribution, gaiaRest, staking, tendermintRpc, tx };
+export {
+  bank,
+  commercio,
+  distribution,
+  gaiaRest,
+  keybase,
+  staking,
+  tendermintRpc,
+  tx,
+};