+++ src/store/validators/__tests__/actions.test.js
@@ -6,6 +6,7 @@ import {
   mockPagination,
   mockPool,
   mockValidator,
+  mockValidatorLogo,
   mockValidatorSets,
 } from '@/__mocks__';
 import actions from '../actions.js';
@@ -124,6 +125,7 @@ describe('store/validators/actions', () => {
     expect(dispatch).toHaveBeenCalledWith('fetchDetail', address);
     expect(dispatch).toHaveBeenCalledWith('fetchDetailDelegations', address);
     expect(dispatch).toHaveBeenCalledWith('fetchPool');
+    expect(dispatch).toHaveBeenCalledWith('setAccount', address);
     expect(commit).toHaveBeenCalledWith('setLoading', false);
     expect(dispatch).not.toHaveBeenCalledWith('fetchTrackedBlocks', lastHeight);
 
@@ -137,17 +139,59 @@ describe('store/validators/actions', () => {
     expect(dispatch).toHaveBeenCalledWith('fetchTrackedBlocks', lastHeight);
   });
 
-  test('if "fetchDetail" action commit "setDetail", and set the error if it is caught', async () => {
+  test('if "fetchDetail" action commit "setDetail" and dispatch "fetchDetailLogo", and set the error if it is caught', async () => {
     const commit = jest.fn();
+    const dispatch = jest.fn();
     const id = 'id';
 
-    await actions.fetchDetail({ commit }, id);
+    await actions.fetchDetail({ commit, dispatch }, id);
 
     expect(commit).toHaveBeenCalledWith('setDetail', mockResponse.data.result);
+    expect(dispatch).toHaveBeenCalledWith(
+      'fetchDetailLogo',
+      mockResponse.data.result.description.identity
+    );
 
     mockError = true;
 
-    await actions.fetchDetail({ commit }, id);
+    await actions.fetchDetail({ commit, dispatch }, id);
+
+    expect(commit).toHaveBeenCalledWith('setError', mockErrorResponse);
+  });
+
+  test('if "fetchDetailLogo" commit "setDetailLogo", and set the logo to "" if error is caught', async () => {
+    const commit = jest.fn();
+    const identity = 'identity';
+
+    await actions.fetchDetailLogo({ commit }, identity);
+
+    for (const item of mockResponse.data.them) {
+      if ('primary' in item['pictures']) {
+        expect(commit).toHaveBeenCalledWith(
+          'setDetailLogo',
+          item['pictures']['primary']['url']
+        );
+      }
+    }
+
+    mockError = true;
+
+    await actions.fetchDetailLogo({ commit }, identity);
+
+    expect(commit).toHaveBeenCalledWith('setDetailLogo', '');
+  });
+
+  test('if "setAccount" commit "setAccount", and set the error if it is caught', () => {
+    const commit = jest.fn();
+    const address = 'address';
+
+    actions.setAccount({ commit }, address);
+
+    expect(commit).toHaveBeenCalledWith('setAccount', mockResponse);
+
+    mockError = true;
+
+    actions.setAccount({ commit }, address);
 
     expect(commit).toHaveBeenCalledWith('setError', mockErrorResponse);
   });
@@ -213,6 +257,23 @@ describe('store/validators/actions', () => {
     expect(commit).toHaveBeenCalledWith('setFilter', filter);
   });
 });
+
+jest.mock('../../../apis/http/keybase.js', () => ({
+  requestValidatorLogo: () => {
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        if (mockError) {
+          reject(mockErrorResponse);
+        }
+
+        mockResponse = {
+          data: mockValidatorLogo(),
+        };
+        resolve(mockResponse);
+      }, 1);
+    });
+  },
+}));
 
 jest.mock('../../../apis/http/staking.js', () => ({
   requestPool: () => {
@@ -295,5 +356,22 @@ jest.mock('../../../apis/http/tendermintRpc.js', () => ({
         resolve(mockResponseValidatorSets);
       }, 1);
     });
+  },
+}));
+
+jest.mock('../../../utils/bech32Manager.js', () => ({
+  decode: () => {
+    if (mockError) {
+      throw mockErrorResponse;
+    }
+    mockResponse = 'decoded_address';
+    return mockResponse;
+  },
+  encode: () => {
+    if (mockError) {
+      throw mockErrorResponse;
+    }
+    mockResponse = 'encoded_address';
+    return mockResponse;
   },
 }));