+++ src/store/validators/actions.js
@@ -1,6 +1,6 @@
-import { staking, tendermintRpc } from '@/apis/http';
+import { keybase, staking, tendermintRpc } from '@/apis/http';
 import { CONFIG, VALIDATORS } from '@/constants';
-import { blocksRequestHelper } from '@/utils';
+import { bech32Manager, blocksRequestHelper } from '@/utils';
 
 export default {
   async initValidatorsList({ commit, dispatch }, lastHeight) {
@@ -59,6 +59,7 @@ export default {
       dispatch('fetchDetail', id),
       dispatch('fetchDetailDelegations', id),
       dispatch('fetchPool'),
+      dispatch('setAccount', id),
     ];
     await Promise.all(requests);
     commit('setLoading', false);
@@ -67,10 +68,45 @@ export default {
     }
   },
 
-  async fetchDetail({ commit }, id) {
+  async fetchDetail({ commit, dispatch }, id) {
     try {
       const response = await staking.requestValidatorsDetailLegacy(id);
       commit('setDetail', response.data.result);
+      const identity = response.data.result.description
+        ? response.data.result.description.identity
+        : '';
+      if (identity) await dispatch('fetchDetailLogo', identity);
+    } catch (error) {
+      commit('setError', error);
+    }
+  },
+
+  async fetchDetailLogo({ commit }, identity) {
+    try {
+      const response = await keybase.requestValidatorLogo(identity);
+      if (!response.data.them || !response.data.them.length) {
+        commit('setDetailLogo', '');
+        return;
+      }
+      for (const item of response.data.them) {
+        if ('primary' in item['pictures']) {
+          commit('setDetailLogo', item['pictures']['primary']['url']);
+          break;
+        }
+      }
+    } catch (error) {
+      commit('setDetailLogo', '');
+    }
+  },
+
+  setAccount({ commit }, address) {
+    try {
+      const hex = bech32Manager.decode(address);
+      const account = bech32Manager.encode(
+        hex,
+        CONFIG.PREFIXES.ACCOUNT.ADDRESS
+      );
+      commit('setAccount', account);
     } catch (error) {
       commit('setError', error);
     }