+++ src/views/validators/__tests__/ValidatorsViewDetail.test.js
@@ -25,8 +25,10 @@ describe('views/validators/ValidatorsViewDetail.vue', () => {
     $store: mockStore,
     $t: (msg) => msg,
   };
+  const props = {
+    id: 'id',
+  };
   const computed = {
-    address: () => 'address',
     lastHeight: () => '1',
     newHeight: () => '2',
   };
@@ -35,18 +37,19 @@ describe('views/validators/ValidatorsViewDetail.vue', () => {
     const wrapper = shallowMount(ValidatorsViewDetail, {
       localVue,
       mocks,
+      propsData: {
+        ...props,
+      },
       computed: {
         ...computed,
         error: () => null,
         isLoading: () => true,
       },
     });
-    await wrapper.setData({ isValid: true });
 
     expect(wrapper.find('[data-test="loading"]').exists()).toBe(true);
     expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
     expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
-    expect(wrapper.find('[data-test="warning"]').exists()).toBe(false);
   });
 
   test('if message error is displayed', async () => {
@@ -54,53 +57,37 @@ describe('views/validators/ValidatorsViewDetail.vue', () => {
     const wrapper = shallowMount(ValidatorsViewDetail, {
       localVue,
       mocks,
+      propsData: {
+        ...props,
+      },
       computed: {
         ...computed,
         error: () => error,
         isLoading: () => false,
       },
     });
-    await wrapper.setData({ isValid: true });
 
     expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
     expect(wrapper.find('[data-test="error"]').exists()).toBe(true);
     expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
-    expect(wrapper.find('[data-test="warning"]').exists()).toBe(false);
   });
 
   test('if content is displayed', async () => {
     const wrapper = shallowMount(ValidatorsViewDetail, {
       localVue,
       mocks,
+      propsData: {
+        ...props,
+      },
       computed: {
         ...computed,
         error: () => null,
         isLoading: () => false,
       },
     });
-    await wrapper.setData({ isValid: true });
 
     expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
     expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
     expect(wrapper.find('[data-test="content"]').exists()).toBe(true);
-    expect(wrapper.find('[data-test="warning"]').exists()).toBe(false);
-  });
-
-  test('if warning message is displayed', async () => {
-    const wrapper = shallowMount(ValidatorsViewDetail, {
-      localVue,
-      mocks,
-      computed: {
-        ...computed,
-        error: () => null,
-        isLoading: () => false,
-      },
-    });
-    await wrapper.setData({ isValid: false });
-
-    expect(wrapper.find('[data-test="loading"]').exists()).toBe(false);
-    expect(wrapper.find('[data-test="error"]').exists()).toBe(false);
-    expect(wrapper.find('[data-test="content"]').exists()).toBe(false);
-    expect(wrapper.find('[data-test="warning"]').exists()).toBe(true);
   });
 });