+++ src/views/validators/detail/ValidatorsViewDetailContentBottomLeft.vue
@@ -31,11 +31,11 @@
 </template>
 
 <script>
-import { mapGetters } from 'vuex';
 import { coinAdapter } from '@/utils';
+import { mapGetters } from 'vuex';
 
 export default {
-  name: 'ValidatorsViewDetailDelegations',
+  name: 'ValidatorsViewDetailContentBottomLeft',
   props: {
     delegations: {
       type: Object,