+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.1.0] - 2022-04-
 
+### Added
+
+- Validator logo on detail view
+
 ### Changed
 
 - Replace icons installation